Update: Module ported to the new format.
