+++ modules/gas/functions/CfgFunctions.hpp
@@ -1,0 +1,13 @@
+class COMPONENT {
+	tag = COMPONENT;
+	class GAS {
+		file = "modules\gas\functions\GAS";
+		class createHotZone {};
+		class fakeMission {};
+		class fireMission {};
+		class gasParticles {};
+		class gasZones {};
+		class jipSend {};
+		class vehicleSiren {};		
+	};
+};
