+++ modules/gas/functions/GAS/fn_jipSend.sqf
@@ -1,0 +1,5 @@
+#include "script_component.hpp"
+
+GVAR(GASLOGICS) apply {
+	["GAS_GASZONES", _x] spawn CBA_fnc_globalEvent;
+};
