+++ modules/gas/functions/GAS/script_component.hpp
@@ -1,0 +1,2 @@
+#define COMPONENT GAS
+#include "..\..\script_component.hpp"
