+++ modules/gas/script_component.hpp
@@ -1,0 +1,2 @@
+#define COMPONENT GAS
+#include "..\..\core\script_macros.hpp"
