+++ modules/gas/sounds.hpp
@@ -1,0 +1,39 @@
+class ACE_Cough_1
+{
+	name = "";
+	sound[] = {"modules\gas\sounds\cough1.ogg",4,1,50};
+	titles[] = {};
+};
+class ACE_Cough_2: ACE_Cough_1
+{
+	sound[] = {"modules\gas\sounds\cough2.ogg",4,1,50};
+};
+class ACE_Cough_3: ACE_Cough_1
+{
+	sound[] = {"modules\gas\sounds\cough3.ogg",4,1,50};
+};
+class ACE_Cough_4: ACE_Cough_1
+{
+	sound[] = {"modules\gas\sounds\cough4.ogg",4,1,50};
+};
+class ACE_Cough_5: ACE_Cough_1
+{
+	sound[] = {"modules\gas\sounds\cough5.ogg",4,1,50};
+};
+class ACE_Cough_H1
+{
+	name = "";
+	sound[] = {"modules\gas\sounds\heavycough1.ogg",4,1,50};
+	titles[] = {};
+};
+class ACE_Cough_H2: ACE_Cough_H1
+{
+	sound[] = {"modules\gas\sounds\heavycough2.ogg",4,1,50};
+};
+
+class ABCA_M42
+{
+	name = "";
+	sound[] = {"modules\gas\sounds\ABCA-M42_1SEC.ogg",4,1,50};
+	titles[] = {};
+};
