+++ modules/gas/sounds/cough1.lip
