+++ modules/gas/sounds/cough2.lip
