+++ modules/gas/sounds/cough2.ogg
