+++ modules/gas/sounds/cough3.lip
