+++ modules/gas/sounds/cough4.lip
