+++ modules/gas/sounds/cough5.lip
