--- modules/gas_script/root.sqf
@@ -1,17 +1,0 @@
-#ifdef preinit
-
-	#include "preInit.sqf"
-
-#endif
-
-#ifdef description_sounds
-
-	#include "sounds.hpp"
-
-#endif
-
-#ifdef framework
-
-	#include "init.sqf"
-
-#endif
